skia: Factor out skiareshape common feature into a dedicated trait
This way we will be able to reuse all its logic for different backends like
OpenGL or vulkan
